myproject/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==3.3.1
Deprecated==1.2.14
Django==5.1.1
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-simple-search==1.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==4.0.1
idna==3.10
inflection==0.5.1
jwcrypto==1.4.2
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2022.7.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
stark-utilities==0.0.5
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
